expose auth via extension client. apply to azure publish

+++ Composer/packages/client/src/pages/botProject/adapters/ABSChannels.tsx
@@ -23,17 +23,9 @@ import TelemetryClient from '../../../telemetry/TelemetryClient';
 import { LoadingSpinner } from '../../../components/LoadingSpinner';
 import { navigateTo } from '../../../utils/navigation';
 import { botDisplayNameState, settingsState } from '../../../recoilModel';
-// import { AuthClient } from '../../../utils/authClient';
-// import { AuthDialog } from '../../../components/Auth/AuthDialog';
-// import {
-//   getTokenFromCache,
-//   getTenantIdFromCache,
-//   isShowAuthDialog,
-//   userShouldProvideTokens,
-// } from '../../../utils/auth';
 import httpClient from '../../../utils/httpUtil';
-import { dispatcherState, primaryTokenState } from '../../../recoilModel';
-import { armScopes } from '../../../constants';
+import { dispatcherState, currentUserState, isAuthenticatedState } from '../../../recoilModel';
+// import { armScopes } from '../../../constants';
 import {
   tableHeaderRow,
   tableRow,
@@ -89,12 +81,10 @@ export enum AzureAPIStatus {
 
 export const ABSChannels: React.FC<RuntimeSettingsProps> = (props) => {
   const { projectId } = props;
-  // const [showAuthDialog, setShowAuthDialog] = useState(false);
   const [currentResource, setCurrentResource] = useState<AzureResourcePointer | undefined>();
   const [channelStatus, setChannelStatus] = useState<AzureChannelsStatus | undefined>();
   const { publishTargets } = useRecoilValue(settingsState(projectId));
   const botDisplayName = useRecoilValue(botDisplayNameState(projectId));
-  // const [token, setToken] = useState<string | undefined>();
   const [availableSubscriptions, setAvailableSubscriptions] = useState<Subscription[]>([]);
   const [publishTargetOptions, setPublishTargetOptions] = useState<IDropdownOption[]>([]);
   const [isLoading, setLoadingStatus] = useState<boolean>(false);
@@ -103,7 +93,8 @@ export const ABSChannels: React.FC<RuntimeSettingsProps> = (props) => {
   const [showTeamsManifestModal, setShowTeamsManifestModal] = useState<boolean>(false);
   const [showTeamsCallOut, setShowTeamsCallOut] = useState<boolean>(false);
   const { setApplicationLevelError, requireUserLogin } = useRecoilValue(dispatcherState);
-  const token = useRecoilValue(primaryTokenState);
+  const currentUser = useRecoilValue(currentUserState);
+  const isAuthenticated = useRecoilValue(isAuthenticatedState);
 
   /* Copied from Azure Publishing extension */
   const getSubscriptions = async (token: string): Promise<Array<Subscription>> => {
@@ -167,18 +158,6 @@ export const ABSChannels: React.FC<RuntimeSettingsProps> = (props) => {
         // TODO: handle case where config does not have a tenant id and we need to fall back to getAccessToken
         requireUserLogin(config.tenantId);
 
-        // let newtoken = '';
-        // if (userShouldProvideTokens()) {
-        //   if (isShowAuthDialog(false)) {
-        //     setShowAuthDialog(true);
-        //   }
-        //   newtoken = getTokenFromCache('accessToken');
-        // } else {
-        //   newtoken = await getTokenInteractively(config.tenantId);
-        // }
-        // setToken(newtoken);
-
-        // if (newtoken) {
         setCurrentResource({
           microsoftAppId: config?.settings?.MicrosoftAppId,
           resourceName: config.botName || config.name,
@@ -186,7 +165,6 @@ export const ABSChannels: React.FC<RuntimeSettingsProps> = (props) => {
           tenantId: config.tenantId,
           subscriptionId: config.subscriptionId,
         });
-        // }
       }
     }
   };
@@ -209,7 +187,7 @@ export const ABSChannels: React.FC<RuntimeSettingsProps> = (props) => {
         }/resourceGroups/${currentResource.resourceGroupName}/providers/Microsoft.BotService/botServices/${
           currentResource.resourceName
         }/channels/${channelId}?api-version=2020-06-02`;
-        await httpClient.get(url, { headers: { Authorization: `Bearer ${token}` } });
+        await httpClient.get(url, { headers: { Authorization: `Bearer ${currentUser.token}` } });
         return {
           enabled: true,
           loading: false,
@@ -290,9 +268,9 @@ export const ABSChannels: React.FC<RuntimeSettingsProps> = (props) => {
             },
           };
       }
-      await httpClient.put(url, data, { headers: { Authorization: `Bearer ${token}` } });
+      await httpClient.put(url, data, { headers: { Authorization: `Bearer ${currentUser.token}` } });
       if (channelId === CHANNELS.TEAMS) {
-        const createResults = await httpClient.get(url, { headers: { Authorization: `Bearer ${token}` } });
+        const createResults = await httpClient.get(url, { headers: { Authorization: `Bearer ${currentUser.token}` } });
         if (!createResults.data?.properties?.properties?.acceptedTerms === true) {
           if (
             await OpenConfirmModal(formatMessage('Microsoft Teams terms and conditions'), null, {
@@ -357,7 +335,7 @@ export const ABSChannels: React.FC<RuntimeSettingsProps> = (props) => {
       }/resourceGroups/${currentResource.resourceGroupName}/providers/Microsoft.BotService/botServices/${
         currentResource.resourceName
       }/channels/${channelId}?api-version=2020-06-02`;
-      await httpClient.delete(url, { headers: { Authorization: `Bearer ${token}` } });
+      await httpClient.delete(url, { headers: { Authorization: `Bearer ${currentUser.token}` } });
 
       // success!!
       setChannelStatus({
@@ -539,16 +517,21 @@ export const ABSChannels: React.FC<RuntimeSettingsProps> = (props) => {
     // reset UI
     setChannelStatus(undefined);
 
-    if (token && currentResource && !currentResource.subscriptionId && !currentResource.alternateSubscriptionId) {
+    if (
+      isAuthenticated &&
+      currentResource &&
+      !currentResource.subscriptionId &&
+      !currentResource.alternateSubscriptionId
+    ) {
       // if we have no subscription id selected, load available subscriptions
 
       // reset the list
       setAvailableSubscriptions([]);
 
       // fetch list of available subscriptions
-      getSubscriptions(token).then((subscriptions) => setAvailableSubscriptions(subscriptions));
+      getSubscriptions(currentUser.token).then((subscriptions) => setAvailableSubscriptions(subscriptions));
     } else if (
-      token &&
+      isAuthenticated &&
       currentResource &&
       (currentResource.subscriptionId || currentResource.alternateSubscriptionId)
     ) {
@@ -564,7 +547,7 @@ export const ABSChannels: React.FC<RuntimeSettingsProps> = (props) => {
       // reset the UI
       setAvailableSubscriptions([]);
     }
-  }, [token, currentResource]);
+  }, [currentUser, isAuthenticated, currentResource]);
 
   const absTableToggle = (key: string) => (
     <Stack horizontal tokens={{ childrenGap: 10 }}>
